refactor(form): refactor controlInput
